+++ src/mk_lib_database_sqlx/src/mk_lib_database_version_schema.rs
@@ -1,5 +1,5 @@
 
-pub async fn mk_lib_database_update_schema(pool: &sqlx::PgPool, version_no: i64)
+pub async fn mk_lib_database_update_schema(pool: &sqlx::PgPool, version_no: i32)
                                            -> Result<bool, sqlx::Error> {
     if version_no < 43 {
         mk_lib_database_version_update(&pool,
@@ -10,7 +10,7 @@ pub async fn mk_lib_database_update_schema(pool: &sqlx::PgPool, version_no: i64)
 
 #[allow(dead_code)]
 pub async fn mk_lib_database_version_update(pool: &sqlx::PgPool,
-                                            version_number: i64)
+                                            version_number: i32)
                                             -> Result<(), sqlx::Error> {
     sqlx::query("update mm_version set mm_version_number = $1")
         .bind(version_number)
